refactor(pkg/cmd/options): port command flags (called options) to options package
